refactor: continue device client architecture cleanup
## Phase 2 – Split core/client.py

Extract responsibilities from the monolithic ThesslaGreenDeviceClient class
into focused mixin modules inside core/:

- core/client_connection.py  – connection lifecycle + transport construction (11 methods)
- core/client_scanner.py     – scanner orchestration + register normalization (4 methods)
- core/client_registers.py   – register groups, IO helpers, write support (7 methods)
- core/client.py             – composition root + __init__ + public API (65% smaller)

core/client.py: 544 → 188 lines.

## Phase 3 – Coordinator proxy cleanup

Replace 6 duplicate method implementations in ThesslaGreenModbusCoordinator
with clean delegation to _device_client:

- get_register_map, _get_client_method → eliminated duplicated logic
- _find_register_name, _process_register_value → eliminated proxy-hop indirection
- _mark_registers_failed, _clear_register_failure → eliminated proxy-hop indirection

Remove 5 now-unused imports (find_register_name_impl, process_register_value_impl,
clear_register_failure_impl, mark_registers_failed_impl, cast).

All 38 coordinator proxy properties retained: required by coordinator submodule
duck-typing and direct test/entity access. Added block documentation comment
explaining retention rationale and future removal path.

## Phase 4 – Dependency direction

Dependency flow verified clean:
  entities/services → coordinator → DeviceClient → coordinator submodules → transport

No circular imports at module level.

## Quality gates

- python -m compileall: clean
- ruff check: All checks passed
- ruff format --check: 432 files already formatted
- tools/check_maintainability.py: Maintainability gate passed
- tools/check_translations.py: All translation keys present
- No merge conflict markers
- No compatibility shims introduced

## No behavioral changes

Modbus behavior, entity IDs, unique IDs, service IDs, register names,
register addresses, and translation keys are all unchanged.
